+++ graalpython/com.oracle.graal.python.test/src/com/oracle/graal/python/test/advance/ExclusionsTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021, 2022, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2020, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -38,51 +38,14 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.oracle.graal.python.builtins.objects.ssl;
+package com.oracle.graal.python.test.advance;
 
-import com.oracle.truffle.api.nodes.ExplodeLoop;
+import org.graalvm.polyglot.Context;
 
-public enum SSLMethod {
-    SSL3(1, SSLProtocol.SSLv3),
-    TLS(2),
-    TLS1(3, SSLProtocol.TLSv1),
-    TLS1_1(4, SSLProtocol.TLSv1_1),
-    TLS1_2(5, SSLProtocol.TLSv1_2),
-    TLS_CLIENT(0x10),
-    TLS_SERVER(0x11);
-
-    private final int pythonId;
-    private final SSLProtocol singleVersion;
-
-    SSLMethod(int pythonId, SSLProtocol singleVersion) {
-        this.pythonId = pythonId;
-        this.singleVersion = singleVersion;
-    }
-
-    SSLMethod(int pythonId) {
-        this.pythonId = pythonId;
-        this.singleVersion = null;
-    }
-
-    public int getPythonId() {
-        return pythonId;
-    }
-
-    public boolean allowsProtocol(SSLProtocol protocol) {
-        return singleVersion == null || singleVersion == protocol;
-    }
-
-    public boolean isSingleVersion() {
-        return singleVersion != null;
-    }
-
-    @ExplodeLoop
-    public static SSLMethod fromPythonId(int pythonId) {
-        for (SSLMethod method : SSLMethod.values()) {
-            if (method.getPythonId() == pythonId) {
-                return method;
-            }
+public class ExclusionsTest {
+    public static void main(String[] args) {
+        try (Context context = Context.create()) {
+            context.eval("python", "print('Hello Python!');");
         }
-        return null;
     }
 }