+++ graalpython/com.oracle.graal.python/src/com/oracle/graal/python/builtins/objects/zipimporter/PZipImporter.java
@@ -37,9 +37,11 @@ import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
+import java.nio.file.NoSuchFileException;
+import java.nio.file.StandardOpenOption;
 import java.util.EnumSet;
 import java.util.zip.ZipEntry;
-import java.util.zip.ZipFile;
+import java.util.zip.ZipInputStream;
 
 import com.oracle.graal.python.builtins.objects.common.HashingStorage;
 import com.oracle.graal.python.builtins.objects.common.HashingStorageNodes.HashingStorageGetIterator;
@@ -50,7 +52,9 @@ import com.oracle.graal.python.builtins.objects.dict.PDict;
 import com.oracle.graal.python.builtins.objects.object.PythonBuiltinObject;
 import com.oracle.graal.python.builtins.objects.str.StringNodes.StringReplaceNode;
 import com.oracle.graal.python.builtins.objects.tuple.PTuple;
+import com.oracle.graal.python.runtime.PythonContext;
 import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
+import com.oracle.truffle.api.TruffleFile;
 import com.oracle.truffle.api.object.Shape;
 import com.oracle.truffle.api.strings.TruffleString;
 import com.oracle.truffle.api.strings.TruffleStringBuilder;
@@ -228,15 +232,52 @@ public class PZipImporter extends PythonBuiltinObject {
      * @throws IOException
      */
     @TruffleBoundary
-    public static TruffleString getCodeFromArchive(TruffleString filenameAndSuffix, TruffleString archive) throws IOException {
-        ZipFile zip = null;
+    public static TruffleString getCodeFromArchive(PythonContext context, TruffleString filenameAndSuffix, TruffleString archive) throws IOException {
+        InputStream fileStream = null;
         try {
-            zip = new ZipFile(archive.toJavaStringUncached());
-            ZipEntry entry = zip.getEntry(filenameAndSuffix.toJavaStringUncached());
-            InputStream in = zip.getInputStream(entry);
+            TruffleFile f = context.getEnv().getPublicTruffleFile(archive.toJavaStringUncached());
+
+            fileStream = f.newInputStream(StandardOpenOption.READ);
+            // This is not correct - it is wrong to read zip files by just skipping to the next
+            // local file header, instead, zip files should always be read from the back, by
+            // scanning for the EOCD record and possibly the EOCD Locator to find a Zip64 EOCD64.
+            // Only the entries and offsets stored in those are valid entries of the Zip file. The
+            // java.util.zip.ZipFile class handles all that, but it only works for java.io.File;
+            // this would break out of our TruffleFile sandbox. Alternative libraries such as zip4j
+            // also do not seem to work properly when just fed input streams. For now, we will just
+            // skip to the first local file header and hope for the best...
+            int offset = 0;
+            byte[] data = fileStream.readNBytes(4096);
+            outer: while (data.length >= 30) { // local file header is 30 bytes minimum
+                int i = 0;
+                for (; i < data.length - 30; i++) {
+                    // zip headers are stored little endian, match the integer value of "PK\03\04",
+                    // the local file header signature
+                    if (data[i] == 'P' && data[i + 1] == 'K' && data[i + 2] == 0x03 && data[i + 3] == 0x04) {
+                        fileStream.close();
+                        fileStream = f.newInputStream(StandardOpenOption.READ);
+                        fileStream.skip(offset + i);
+                        break outer;
+                    }
+                }
+                offset += 4096;
+                data = fileStream.readNBytes(4096);
+            }
+
+            ZipInputStream zis = new ZipInputStream(fileStream);
+            ZipEntry entry;
+            String filenameAndSuffixStr = filenameAndSuffix.toJavaStringUncached();
+            while ((entry = zis.getNextEntry()) != null) {
+                if (entry.getName().equals(filenameAndSuffixStr)) {
+                    break;
+                }
+            }
+            if (entry == null) {
+                throw new NoSuchFileException(filenameAndSuffixStr);
+            }
 
             // reading the file should be done better?
-            BufferedReader reader = new BufferedReader(new InputStreamReader(in));
+            BufferedReader reader = new BufferedReader(new InputStreamReader(zis));
             int size = (int) entry.getSize();
             if (size < 0) {
                 size = (int) entry.getCompressedSize();
@@ -253,9 +294,9 @@ public class PZipImporter extends PythonBuiltinObject {
         } catch (IOException e) {
             throw e;
         } finally {
-            if (zip != null) {
+            if (fileStream != null) {
                 try {
-                    zip.close();
+                    fileStream.close();
                 } catch (IOException e) {
                     // just ignore it.
                 }
@@ -320,7 +361,7 @@ public class PZipImporter extends PythonBuiltinObject {
     }
 
     @TruffleBoundary
-    protected final ModuleCodeData getModuleCode(TruffleString fullname) throws IOException {
+    protected final ModuleCodeData getModuleCode(PythonContext context, TruffleString fullname) throws IOException {
         TruffleString path = makeFilename(fullname);
         TruffleString fullPath = makePackagePath(fullname);
 
@@ -338,7 +379,7 @@ public class PZipImporter extends PythonBuiltinObject {
 
             TruffleString code;
             try {
-                code = getCodeFromArchive(searchPath, archive);
+                code = getCodeFromArchive(context, searchPath, archive);
             } catch (IOException e) {
                 throw new IOException("Can not read code from " + makePackagePath(searchPath), e);
             }