+++ graalpython/com.oracle.graal.python/src/com/oracle/graal/python/runtime/PythonContext.java
@@ -1516,9 +1516,11 @@ public final class PythonContext extends Python3Core {
             initializeHashSecret();
         }
         setIntMaxStrDigits(getOption(PythonOptions.IntMaxStrDigits));
-        nativeZlib = NFIZlibSupport.createNative(this, "");
-        nativeBz2lib = NFIBz2Support.createNative(this, "");
-        nativeLZMA = NFILZMASupport.createNative(this, "");
+        if (!PythonOptions.WITHOUT_COMPRESSION_LIBRARIES) {
+            nativeZlib = NFIZlibSupport.createNative(this, "");
+            nativeBz2lib = NFIBz2Support.createNative(this, "");
+            nativeLZMA = NFILZMASupport.createNative(this, "");
+        }
 
         mainModule = factory().createPythonModule(T___MAIN__);
         mainModule.setAttribute(T___BUILTINS__, getBuiltins());
@@ -1569,7 +1571,12 @@ public final class PythonContext extends Python3Core {
         PosixSupport result;
         // The resources field will be removed once all posix builtins go through PosixSupport
         TruffleString.EqualNode eqNode = TruffleString.EqualNode.getUncached();
-        if (eqNode.execute(T_JAVA, option, TS_ENCODING)) {
+        boolean selectedJavaBackend = eqNode.execute(T_JAVA, option, TS_ENCODING);
+        if (PythonOptions.WITHOUT_NATIVE_POSIX || selectedJavaBackend) {
+            if (!selectedJavaBackend) {
+                writeWarning("Native Posix backend selected, but it was excluded from the runtime, " +
+                                "switching to Java backend.");
+            }
             result = new EmulatedPosixSupport(this);
         } else if (eqNode.execute(T_NATIVE, option, TS_ENCODING) || eqNode.execute(T_LLVM_LANGUAGE, option, TS_ENCODING)) {
             if (ImageInfo.inImageBuildtimeCode()) {